testovací raspberry/bch-tester-core2/requirements.txt
appdirs>=1.0
pyserial>=3.0
bcf>=0.17
httplib2
google-api-python-client
oauth2client

